Talent Intelligence & Workforce Strategy Agent/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9

langgraph==0.0.45
langchain==0.1.14

neo4j==5.18.0

pydantic==2.6.4
pydantic-settings==2.2.1

reportlab==4.1.0

python-dotenv==1.0.1
requests==2.31.0

jinja2==3.1.3